working in progress

+++ src/init/init.h
@@ -6,7 +6,7 @@
 /*   By: kvebers <kvebers@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/23 12:05:44 by kvebers           #+#    #+#             */
-/*   Updated: 2023/05/27 10:14:48 by kvebers          ###   ########.fr       */
+/*   Updated: 2023/05/29 13:42:19 by kvebers          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void		test_textures(t_params *params);
 //VECTORS.c
 t_vector	vector_estimation(int speed, int angle);
 //MOVE_VALID.c
-void	check_valid_move(t_params *params, float posx, float posy);
-void	check_valid_move1(t_params *params, float posx, float posy);
+void	check_valid_move(t_params *params, double posx, double posy);
+void	check_valid_move1(t_params *params, double posx, double posy);
 int     check_cordinates(t_params *params, int x, int y);
 #endif
